% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stats.R
\name{barplotly_statistics}
\alias{barplotly_statistics}
\title{plotly barplot}
\usage{
barplotly_statistics(
  bike_data = parsed_bike_data[[3]],
  datelim = c(as.Date("2020-01-01", "\%Y-\%m-\%d"), as.Date("2022-10-31", "\%Y-\%m-\%d")),
  plot_colors = default_colors(default = F)
)
}
\value{
plotly barplot
}
\description{
pplotly barplot
}
\examples{
eg = barplotly_statistics();
}
