% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stats.R
\name{default_colors}
\alias{default_colors}
\title{default colors}
\usage{
default_colors(default = T, size = 55)
}
\arguments{
\item{default}{logical. R colors or from Wes Anderson movies.}

\item{size}{int. number of colors you need.}
}
\value{
vector of colors
}
\description{
This generate a vector of colors
}
\examples{
eg = default_colors();
}
