% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stats.R
\name{loess_plotly}
\alias{loess_plotly}
\title{loess_plotly}
\usage{
loess_plotly(
  data = dummy_data(),
  stations = "Boyer / Everett",
  datelim = c(as.Date("2019-01-01", "\%Y-\%m-\%d"), as.Date("2022-12-31", "\%Y-\%m-\%d")),
  moyenne_globale = FALSE,
  plot_colors = default_colors(default = F, size = nrow(data[[3]])),
  language = 0
)
}
\arguments{
\item{data}{list. input data from parse_bike_data(). A list of size 3.}

\item{stations}{str. stations}

\item{datelim}{date. dates}

\item{moyenne_globale}{logical.}

\item{plot_colors}{a vector of colors to look pretty}

\item{moyenne_mobile}{logical.}
}
\value{
plotly line plot
}
\description{
create a line plot of smoothed average
}
\examples{
fig = loess_plotly()
}
