% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{parse_bike_data}
\alias{parse_bike_data}
\title{parse bike data}
\usage{
parse_bike_data(path = "VeloMtl/data/")
}
\arguments{
\item{path}{Where is the data?}
}
\value{
parsed bike data
}
\description{
parse bike data
}
\examples{
parsed_bike_data = parse_bike_data()
}
