% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{read_bike_data}
\alias{read_bike_data}
\title{read data}
\usage{
read_bike_data(path = "data/", recalculate = F)
}
\arguments{
\item{recalculate}{logical. If you want to recalculate from the raw files..}

\item{path.}{Where is the data?}
}
\value{
data
}
\description{
read data
}
\examples{
eg = read_bike_data();
}
