% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stats.R
\name{scatter_stats_plotly}
\alias{scatter_stats_plotly}
\title{scatter_stats_plotly}
\usage{
scatter_stats_plotly(
  data = parsed_bike_data,
  stations = "Brébeuf / Rachel",
  datelim = c(as.Date("2020-01-01", "\%Y-\%m-\%d"), as.Date("2022-10-31", "\%Y-\%m-\%d")),
  moyenne_globale = FALSE,
  add_similar = FALSE,
  moyenne_mobile = FALSE,
  plot_colors = default_colors(default = F)
)
}
\value{
plotly
}
\description{
scatter_stats_plotly
}
\examples{
eg = scatter_stats_plotly();
}
