% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stats.R
\name{barplotly_statistics}
\alias{barplotly_statistics}
\title{barplotly_statistics}
\usage{
barplotly_statistics(
  data = dummy_data(),
  datelim = "",
  plot_colors = default_colors(default = F),
  language = 0
)
}
\arguments{
\item{data}{list. input data from parse_bike_data(). A list of size 3.}

\item{datelim}{date. dates}

\item{plot_colors}{a vector of colors to look pretty.}
}
\value{
plotly bar plot
}
\description{
Create a bar plot for all stations
}
\examples{
fig = barplotly_statistics(data =list("a","b",data.frame(Nom=letters[1:10],sum=sample(1:10),ID = seq(1001,1010))) )
}
