% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stats.R
\name{loess_plotly}
\alias{loess_plotly}
\title{loess_plotly}
\usage{
loess_plotly(
  data = parsed_bike_data,
  stations = "Brébeuf / Rachel",
  datelim = c(as.Date("2020-01-01", "\%Y-\%m-\%d"), as.Date("2022-10-31", "\%Y-\%m-\%d")),
  add_similar = FALSE,
  moyenne_globale = FALSE,
  moyenne_mobile = FALSE,
  plot_colors = default_colors(default = F)
)
}
\value{
plotly
}
\description{
loess_plotly
}
\examples{
eg = loess_plotly();
}
