% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{parse_bike_data}
\alias{parse_bike_data}
\title{parse bike data}
\usage{
parse_bike_data(path = "data/", recalculate = F)
}
\arguments{
\item{path}{Where is the data?}

\item{recalculate}{logical. If you want to recalculate from the raw files.}
}
\value{
parsed bike data
}
\description{
parse bike data
}
\examples{
eg = parse_bike_data();
}
