% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stats.R
\name{scatter_stats_plotly}
\alias{scatter_stats_plotly}
\title{scatter_stats_plotly}
\usage{
scatter_stats_plotly(
  data = parsed_bike_data,
  stations = "Brébeuf / Rachel",
  datelim = c(as.Date("2020-01-01", "\%Y-\%m-\%d"), as.Date("2022-10-31", "\%Y-\%m-\%d")),
  moyenne_globale = FALSE,
  moyenne_mobile = FALSE,
  plot_colors = default_colors(default = F)
)
}
\arguments{
\item{data}{input data from parse_bike_data(). A list of size 3.}

\item{stations}{str. stations.}

\item{datelim}{date. dates.}

\item{moyenne_globale}{logical.}

\item{moyenne_mobile}{logical.}

\item{plot_colors}{a vector of colors to look pretty}
}
\value{
a scatter plotly plot
}
\description{
create a scatter plot of counts per day
}
\examples{
eg = scatter_stats_plotly();
}
